healthcare-assistant-project/backend/requirements.txt

fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.21
langchain-gemini==0.0.6
pinecone-client==3.0.2
langchain-pinecone==0.0.2
requests==2.31.0
pydantic==2.6.3
python-multipart==0.0.6